+++ DependencyInjection/Compiler/RegisterMappingPass.php
@@ -24,19 +24,11 @@ class RegisterMappingPass implements CompilerPassInterface
         $name = empty($storage['object_manager']) ? 'default' : $storage['object_manager'];
 
         $ormDriverId     = sprintf('doctrine.orm.%s_metadata_driver', $name);
-        $mongodbDriverId = sprintf('doctrine_mongodb.odm.%s_metadata_driver', $name);
 
         if (StorageInterface::STORAGE_ORM == $storage['type'] && $container->hasDefinition($ormDriverId)) {
             $container->getDefinition($ormDriverId)->addMethodCall(
                 'addDriver',
                 array(new Reference('lexik_translation.orm.metadata.xml'), 'Lexik\Bundle\TranslationBundle\Model')
-            );
-        }
-
-        if (StorageInterface::STORAGE_MONGODB == $storage['type'] && $container->hasDefinition($mongodbDriverId)) {
-            $container->getDefinition($mongodbDriverId)->addMethodCall(
-                'addDriver',
-                array(new Reference('lexik_translation.mongodb.metadata.xml'), 'Lexik\Bundle\TranslationBundle\Model')
             );
         }
     }