+++ Manager/TransUnitManager.php
@@ -4,7 +4,6 @@ namespace Lexik\Bundle\TranslationBundle\Manager;
 
 use Lexik\Bundle\TranslationBundle\Model\Translation;
 use Lexik\Bundle\TranslationBundle\Storage\StorageInterface;
-use Lexik\Bundle\TranslationBundle\Storage\PropelStorage;
 
 /**
  * Class to manage TransUnit entities or documents.
@@ -145,10 +144,6 @@ class TransUnitManager implements TransUnitManagerInterface
             $translation->setContent($content);
         }
 
-        if (null !== $translation && $this->storage instanceof PropelStorage) {
-            $this->storage->persist($translation);
-        }
-
         if ($flush) {
             $this->storage->flush();
         }
@@ -173,9 +168,6 @@ class TransUnitManager implements TransUnitManagerInterface
 
                     $contentUpdated = ($translation->getContent() != $originalContent);
 
-                    if ($this->storage instanceof PropelStorage) {
-                        $this->storage->persist($transUnit);
-                    }
                 } else {
                     //We need to get a proper file for this translation
                     $file = $this->getTranslationFile($transUnit, $locale);