+++ Translation/Exporter/JsonExporter.php
@@ -27,7 +27,7 @@ class JsonExporter implements ExporterInterface
      */
     public function export($file, $translations)
     {
-        $bytes = file_put_contents($file, json_encode($this->hierarchicalFormat ? $this->hierarchicalFormat($translations) : $translations, JSON_PRETTY_PRINT));
+        $bytes = file_put_contents($file, json_encode($this->hierarchicalFormat ? $this->hierarchicalFormat($translations) : $translations, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES));
 
         return ($bytes !== false);
     }